@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 
-require("../src/cli");
+import "../src/cli.js";
